requirements.txt
sqlmodel==0.0.31
sqlalchemy>=2.0
asyncpg
fastapi
uvicorn[standard]
python-dotenv
httpx
pytest
pytest-asyncio
